+++ lib/wipco/auth.rb
@@ -1,7 +1,0 @@
-require "wipco"
-
-class Wipco::Auth
-  def self.api_key
-    ENV['WIPCO_API_KEY']
-  end
-end