+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - Renamed command from wipco to wip
+- Renamed gem from wipco to wip-cli
 
 ## [1.0.0] - 2020-11-10